Bump back to 1.11
Reverted and renamed holi.core/abs so that clojure.core/abs would
not conlfict. I'm not 100% sure where the conflict is, but in some way
if I use 1.11 and try to use abs in holi.core, the tests stall.

So decided to just rename the abs in holi.core as absolute