Make 'calendars-generated' THE place for datelists
Problem here was that when running tests it worked because it was
reading datelist files from resources/calendars-generated, but when it
got built the files went to the root dir and then it wouldn't work.

Now the build puts the files in 'calendars-generated' too. Now just have
to remove all that duplication of 'calendars-generated' everywhere...